Initial TransferDao and JdbcTransferDao

Approved-by: DJ Akinbode

+++ tenmo-server/src/main/java/com/techelevator/tenmo/model/Transfer.java
@@ -4,29 +4,29 @@ import java.math.BigDecimal;
 
 public class Transfer {
 
-    private long transferId;
-    private long transferTypeId;
-    private long transferStatusId;
-    private long accountFrom;
-    private long accountTo;
+    private Long transferId;
+    private int transferTypeId;
+    private int transferStatusId;
+    private Long accountFrom;
+    private Long accountTo;
     private BigDecimal amount;
 
     public Transfer() { }
 
-    public Transfer(long transferId, long transferTypeId, long transferStatusId, long accountFrom, long accountTo, BigDecimal amount ){
+    public Transfer(Long transferId, int transferTypeId, int transferStatusId, Long accountFrom, Long accountTo, BigDecimal amount) {
         this.transferId = transferId;
         this.transferTypeId = transferTypeId;
         this.transferStatusId = transferStatusId;
         this.accountFrom = accountFrom;
-        this.accountTo=accountTo;
-        this.amount=amount;
+        this.accountTo = accountTo;
+        this.amount = amount;
     }
 
     public long getTransferId() {
         return transferId;
     }
 
-    public void setTransferId(long transferId) {
+    public void setTransferId(Long transferId) {
         this.transferId = transferId;
     }
 
@@ -34,7 +34,7 @@ public class Transfer {
         return transferTypeId;
     }
 
-    public void setTransferTypeId(long transferTypeId) {
+    public void setTransferTypeId(int transferTypeId) {
         this.transferTypeId = transferTypeId;
     }
 
@@ -42,7 +42,7 @@ public class Transfer {
         return transferStatusId;
     }
 
-    public void setTransferStatusId(long transferStatusId) {
+    public void setTransferStatusId(int transferStatusId) {
         this.transferStatusId = transferStatusId;
     }
 
@@ -50,7 +50,7 @@ public class Transfer {
         return accountFrom;
     }
 
-    public void setAccountFrom(long accountFrom) {
+    public void setAccountFrom(Long accountFrom) {
         this.accountFrom = accountFrom;
     }
 
@@ -58,7 +58,7 @@ public class Transfer {
         return accountTo;
     }
 
-    public void setAccountTo(long accountTo) {
+    public void setAccountTo(Long accountTo) {
         this.accountTo = accountTo;
     }
